Rename sink.Resource to sink.Sink
Resource is too generic and Sink seems like a more apt name.
Also, fixes the indenting in a bunch of tests to be more consistent
with they style from pipelines.
